Manager creation validation methods added to UserServiceImpl and IUserService
